Move to TextFilters folder & add doc
